move deprecated files
